requirements.txt
keras==2.10.0
matplotlib==3.5.3
numpy==1.22.4
pandas==1.4.2
python-dotenv==1.0.0
scikit_learn==1.2.2
